Mostly work on the client
Added some fields to critter for future work: version, which we should
pass and environment_hash, which we will need to calculate in a very
clever way. The client doesn't know which profile critterding is using
so it can't get this information.

This can't be solved easily without asking critterding itself which
config it is using. Ideally we should build the client into critterding,
that will take a huge amount of copy/extract/hash/copy-back complexity
out of the system. That's why I'm not bothering trying to figure out the
config profile and hash it and send it right now.